+++ effectful-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Make storage references in the environment immutable.
 * Remove `checkSizeEnv` and `forkEnv` from
   `Effectful.Dispatch.Static.Primitive`.
+* Add internal versioning of effects to prevent leakage of `unsafeCoerce`.
 
 # effectful-core-1.2.0.0 (2022-07-28)
 * Change `SuffixOf` to `SharedSuffix` and make it behave as advertised.